Revert "Minor Code improvements"
This reverts commit c25cba140a0e728f9931d48851d0b6f82bf1daf6.

+++ django_fine_uploader_s3/views.py
@@ -1,12 +1,10 @@
 import base64
 import hashlib
 import hmac
+import uuid
 import json
 import mimetypes
-import uuid
 
-import boto
-from boto.s3.connection import Key, S3Connection
 from django.conf import settings
 from django.http import HttpResponse
 from django.views.decorators.csrf import csrf_exempt
@@ -14,6 +12,9 @@ from rest_framework.decorators import permission_classes
 from rest_framework.permissions import IsAuthenticated
 
 AWS_MAX_SIZE = 15000000000
+
+import boto
+from boto.s3.connection import Key, S3Connection
 
 
 @csrf_exempt
@@ -38,9 +39,9 @@ def success_redirect_endpoint(request):
 
 
 def handle_POST(request):
-    """ Handle S3 uploader POST requests here. For files <=5MiB this is a
-    simple request to sign the policy document. For files >5MiB this is a
-    request to sign the headers to start a multipart encoded request.
+    """ Handle S3 uploader POST requests here. For files <=5MiB this is a simple
+    request to sign the policy document. For files >5MiB this is a request
+    to sign the headers to start a multipart encoded request.
     """
     if request.POST.get('success', None):
         return make_response(200)
@@ -69,22 +70,24 @@ def handle_DELETE(request):
         boto.set_stream_logger('boto')
         S3 = S3Connection(aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
                           aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY)
-    except ImportError:
+    except ImportError as e:
         print("Could not import boto, the Amazon SDK for Python.")
         print("Deleting files will not work.")
         print("Install boto with")
         print("$ pip install boto")
-        return make_response(500)
 
-    bucket_name = request.POST.get('bucket')
-    key_name = request.POST.get('key')
-    try:
-        aws_bucket = S3.get_bucket(bucket_name, validate=False)
-        aws_key = Key(aws_bucket, key_name)
-        aws_key.delete()
-        return make_response(200)
-    except Exception as err:
-        print(err)
+    if boto:
+        bucket_name = request.POST.get('bucket')
+        key_name = request.POST.get('key')
+        try:
+            aws_bucket = S3.get_bucket(bucket_name, validate=False)
+            aws_key = Key(aws_bucket, key_name)
+            aws_key.delete()
+            return make_response(200)
+        except Exception as err:
+            print(err)
+            return make_response(500)
+    else:
         return make_response(500)
 
 
@@ -107,14 +110,12 @@ def is_valid_policy(policy_document):
     bucket = ''
     parsed_max_size = 0
     for condition in policy_document['conditions']:
-        content_len_range = 'content-length-range'
-        if isinstance(condition, list) and condition[0] == content_len_range:
+        if isinstance(condition, list) and condition[0] == 'content-length-range':
             parsed_max_size = condition[2]
         else:
             if condition.get('bucket', None):
                 bucket = condition['bucket']
-    return bucket == settings.AWS_STORAGE_BUCKET_NAME and int(
-        parsed_max_size) == AWS_MAX_SIZE
+    return bucket == settings.AWS_STORAGE_BUCKET_NAME and int(parsed_max_size) == AWS_MAX_SIZE
 
 
 def sign_policy_document(policy_document):
@@ -122,9 +123,7 @@ def sign_policy_document(policy_document):
     http://aws.amazon.com/articles/1434/#signyours3postform
     """
     policy = base64.b64encode(json.dumps(policy_document))
-    signature = base64.b64encode(
-        hmac.new(settings.AWS_SECRET_ACCESS_KEY, policy,
-                 hashlib.sha1).digest())
+    signature = base64.b64encode(hmac.new(settings.AWS_SECRET_ACCESS_KEY, policy, hashlib.sha1).digest())
     return {
         'policy': policy,
         'signature': signature
@@ -134,9 +133,7 @@ def sign_policy_document(policy_document):
 def sign_headers(headers):
     """ Sign and return the headers for a chunked upload. """
     return {
-        'signature': base64.b64encode(
-            hmac.new(settings.AWS_SECRET_ACCESS_KEY, headers,
-                     hashlib.sha1).digest())
+        'signature': base64.b64encode(hmac.new(settings.AWS_SECRET_ACCESS_KEY, headers, hashlib.sha1).digest())
     }
 
 
@@ -146,8 +143,7 @@ def sign_s3_upload(request):
     folder_name = request.GET["folderName"]
     object_name = str(uuid.uuid4()) + "-" + object_name
     key_name = folder_name + "/" + object_name
-    content_type = request.GET.get("contentType",
-                                   mimetypes.guess_type(object_name)[0])
+    content_type = request.GET.get("contentType", mimetypes.guess_type(object_name)[0])
     bucket_name = settings.AWS_STORAGE_BUCKET_NAME
 
     import boto3
@@ -155,8 +151,8 @@ def sign_s3_upload(request):
 
     # Get the service client with sigv4 configured
     s3 = boto3.client('s3', aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
-                      aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-                      config=Config(signature_version='s3v4'))
+                      aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY, config=Config(signature_version='s3v4'))
+
 
     # Generate the URL to get 'key-name' from 'bucket-name'
     signed_url = s3.generate_presigned_url(
@@ -170,8 +166,7 @@ def sign_s3_upload(request):
     )
 
     cloud_front = getattr(settings, 'AWS_CLOUDFRONT_DOMAIN', None)
-    cloud_front_url = "https://%s.s3.amazonaws.com/%s" % (
-        bucket_name, key_name)
+    cloud_front_url = "https://%s.s3.amazonaws.com/%s" % (bucket_name, key_name)
     if cloud_front:
         cloud_front_url = "https://%s/%s" % (cloud_front, key_name)
     response = {
